Add status bar entry for forwarded ports Fixes microsoft/vscode-remote-release#3825
