+++ src/vs/workbench/contrib/remote/browser/remote.ts
@@ -17,9 +17,9 @@ import { IThemeService } from 'vs/platform/theme/common/themeService';
 import { IContextMenuService } from 'vs/platform/contextview/browser/contextView';
 import { IExtensionService } from 'vs/workbench/services/extensions/common/extensions';
 import { FilterViewPaneContainer } from 'vs/workbench/browser/parts/views/viewsViewlet';
-import { ForwardedPortsView, VIEWLET_ID } from 'vs/workbench/contrib/remote/browser/remoteExplorer';
+import { AutomaticPortForwarding, ForwardedPortsView, VIEWLET_ID } from 'vs/workbench/contrib/remote/browser/remoteExplorer';
 import { IContextKeyService } from 'vs/platform/contextkey/common/contextkey';
-import { IViewDescriptor, IViewsRegistry, Extensions, ViewContainerLocation, IViewContainersRegistry, IViewDescriptorService, IViewsService } from 'vs/workbench/common/views';
+import { IViewDescriptor, IViewsRegistry, Extensions, ViewContainerLocation, IViewContainersRegistry, IViewDescriptorService } from 'vs/workbench/common/views';
 import { Registry } from 'vs/platform/registry/common/platform';
 import { IExtensionDescription } from 'vs/platform/extensions/common/extensions';
 import { IOpenerService } from 'vs/platform/opener/common/opener';
@@ -37,14 +37,13 @@ import { IDialogService } from 'vs/platform/dialogs/common/dialogs';
 import { ReconnectionWaitEvent, PersistentConnectionEventType } from 'vs/platform/remote/common/remoteAgentConnection';
 import Severity from 'vs/base/common/severity';
 import { ReloadWindowAction } from 'vs/workbench/browser/actions/windowActions';
-import { Disposable, IDisposable } from 'vs/base/common/lifecycle';
+import { IDisposable } from 'vs/base/common/lifecycle';
 import { LifecyclePhase } from 'vs/platform/lifecycle/common/lifecycle';
 import { SwitchRemoteViewItem, SwitchRemoteAction } from 'vs/workbench/contrib/remote/browser/explorerViewItems';
 import { Action, IActionViewItem, IAction } from 'vs/base/common/actions';
 import { isStringArray } from 'vs/base/common/types';
-import { IRemoteExplorerService, MakeAddress, mapHasTunnelLocalhostOrAllInterfaces } from 'vs/workbench/services/remote/common/remoteExplorerService';
+import { IRemoteExplorerService } from 'vs/workbench/services/remote/common/remoteExplorerService';
 import { IWorkbenchEnvironmentService } from 'vs/workbench/services/environment/common/environmentService';
-import { forwardedPortsViewEnabled, OpenPortInBrowserAction } from 'vs/workbench/contrib/remote/browser/tunnelView';
 import { ViewPane, IViewPaneOptions } from 'vs/workbench/browser/parts/views/viewPaneContainer';
 import { IListVirtualDelegate } from 'vs/base/browser/ui/list/list';
 import { ITreeRenderer, ITreeNode, IAsyncDataSource } from 'vs/base/browser/ui/tree/tree';
@@ -56,9 +55,6 @@ import { SyncDescriptor } from 'vs/platform/instantiation/common/descriptors';
 import { RemoteStatusIndicator } from 'vs/workbench/contrib/remote/browser/remoteIndicator';
 import { inQuickPickContextKeyValue } from 'vs/workbench/browser/quickaccess';
 import { Codicon, registerIcon } from 'vs/base/common/codicons';
-import { INotificationService, IPromptChoice } from 'vs/platform/notification/common/notification';
-import { UrlFinder } from 'vs/workbench/contrib/remote/browser/urlFinder';
-import { ITerminalService } from 'vs/workbench/contrib/terminal/browser/terminal';
 
 export interface HelpInformation {
 	extensionDescription: IExtensionDescription;
@@ -823,91 +819,6 @@ class RemoteAgentConnectionStatusListener implements IWorkbenchContribution {
 						break;
 				}
 			});
-		}
-	}
-}
-
-class AutomaticPortForwarding extends Disposable implements IWorkbenchContribution {
-	private contextServiceListener?: IDisposable;
-	private urlFinder?: UrlFinder;
-	private static AUTO_FORWARD_SETTING = 'remote.autoForwardPorts';
-
-	constructor(
-		@ITerminalService private readonly terminalService: ITerminalService,
-		@INotificationService private readonly notificationService: INotificationService,
-		@IOpenerService private readonly openerService: IOpenerService,
-		@IViewsService private readonly viewsService: IViewsService,
-		@IRemoteExplorerService private readonly remoteExplorerService: IRemoteExplorerService,
-		@IWorkbenchEnvironmentService private readonly environmentService: IWorkbenchEnvironmentService,
-		@IContextKeyService private readonly contextKeyService: IContextKeyService,
-		@IConfigurationService private readonly configurationService: IConfigurationService
-	) {
-		super();
-		this._register(configurationService.onDidChangeConfiguration((e) => {
-			if (e.affectsConfiguration(AutomaticPortForwarding.AUTO_FORWARD_SETTING)) {
-				this.tryStartStopUrlFinder();
-			}
-		}));
-
-		if (this.environmentService.remoteAuthority) {
-			this.contextServiceListener = this._register(this.contextKeyService.onDidChangeContext(e => {
-				if (e.affectsSome(new Set(forwardedPortsViewEnabled.keys()))) {
-					this.tryStartStopUrlFinder();
-				}
-			}));
-			this.tryStartStopUrlFinder();
-		}
-	}
-
-	private tryStartStopUrlFinder() {
-		if (this.configurationService.getValue(AutomaticPortForwarding.AUTO_FORWARD_SETTING)) {
-			this.startUrlFinder();
-		} else {
-			this.stopUrlFinder();
-		}
-	}
-
-	private startUrlFinder() {
-		if (!this.urlFinder && !forwardedPortsViewEnabled.getValue(this.contextKeyService)) {
-			return;
-		}
-		if (this.contextServiceListener) {
-			this.contextServiceListener.dispose();
-		}
-		this.urlFinder = this._register(new UrlFinder(this.terminalService));
-		this._register(this.urlFinder.onDidMatchLocalUrl(async (localUrl) => {
-			if (mapHasTunnelLocalhostOrAllInterfaces(this.remoteExplorerService.tunnelModel.forwarded, localUrl.host, localUrl.port)) {
-				return;
-			}
-			const forwarded = await this.remoteExplorerService.forward(localUrl);
-			if (forwarded) {
-				const address = MakeAddress(forwarded.tunnelRemoteHost, forwarded.tunnelRemotePort);
-				const message = nls.localize('remote.tunnelsView.automaticForward', "{0} from the remote has been forwarded to {1} locally.",
-					address, forwarded.localAddress);
-				const browserChoice: IPromptChoice = {
-					label: OpenPortInBrowserAction.LABEL,
-					run: () => OpenPortInBrowserAction.run(this.remoteExplorerService.tunnelModel, this.openerService, address)
-				};
-				const showChoice: IPromptChoice = {
-					label: nls.localize('remote.tunnelsView.showView', "Show Forwarded Ports"),
-					run: () => {
-						const remoteAuthority = this.environmentService.remoteAuthority;
-						const explorerType: string[] | undefined = remoteAuthority ? [remoteAuthority.split('+')[0]] : undefined;
-						if (explorerType) {
-							this.remoteExplorerService.targetType = explorerType;
-						}
-						this.viewsService.openViewContainer(VIEWLET_ID);
-					}
-				};
-				this.notificationService.prompt(Severity.Info, message, [browserChoice, showChoice], { neverShowAgain: { id: 'remote.tunnelsView.autoForwardNeverShow', isSecondary: true } });
-			}
-		}));
-	}
-
-	private stopUrlFinder() {
-		if (this.urlFinder) {
-			this.urlFinder.dispose();
-			this.urlFinder = undefined;
 		}
 	}
 }
